refactor(core): Phase 0 de-fusion — extract service spawns + store init, drop dead dispatch tier
Pure motion, no behavior change:
- Delete always-None tier-3 dispatch shim (src/rpc/dispatch.rs) and its call site
- Convert src/core/runtime.rs to a directory module
- Extract workspace-bound store init into runtime::context::init_stores
- Extract the four background service spawns into runtime::services::spawn_*

Groundwork for the CoreBuilder/ServiceSet composition in Phase 1.
See docs/plans/pluggable-core/phase-0-defusion.md.
