+++ src/core/runtime/mod.rs
@@ -1,4 +1,13 @@
-//! Shared tokio runtime tuning constants.
+//! Core runtime composition: bootstrap context, background services, and
+//! shared tokio tuning constants.
+//!
+//! This module is the seam that separates *initialization* (workspace-bound
+//! store setup — [`context`]) from *background services* (cron, channels,
+//! heartbeat, update scheduler — [`services`]) so alternate hosts can compose
+//! them without going through the monolithic `run_server_inner`. See
+//! `docs/plans/pluggable-core/` for the full plan.
+//!
+//! ## Shared tokio runtime tuning constants
 //!
 //! A single agent turn is a very large async state machine (system prompt +
 //! hundreds of tool specs + the nested provider/tool loop), and delegating
@@ -14,3 +23,6 @@
 //! in sync; downstream call sites should set `.thread_stack_size(AGENT_WORKER_STACK_BYTES)`
 //! on every multi-thread runtime that may host an agent turn.
 pub const AGENT_WORKER_STACK_BYTES: usize = 16 * 1024 * 1024;
+
+pub mod context;
+pub mod services;
